# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AudioInput {
  /** Base64-encoded audio data */
  audioBase64: string;
  /** audio/webm, audio/wav, audio/mp4, etc. */
  mimeType: string;
}

export interface TTSInput {
  /** The text to synthesize into speech */
  text: string;
  /** Voice ID (e.g. autumn, diana, austin, etc.) */
  voice?: string;
  /** TTS model to use */
  model?: string;
}

export interface ChatInput {
  message: string;
  /** @nullable */
  conversationId?: number | null;
  /** @nullable */
  model?: string | null;
  /**
     * groq | nvidia
     * @nullable
     */
  provider?: string | null;
  /**
     * Optional base64 image data for multimodal vision models
     * @nullable
     */
  imageBase64?: string | null;
}

export interface ChatResponse {
  reply: string;
  conversationId: number;
  messageId: number;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  tokensUsed?: number | null;
  /** @nullable */
  toolsUsed?: string[] | null;
}

export interface TranscribeResponse {
  text: string;
  /** @nullable */
  confidence?: number | null;
}

export interface Settings {
  id: number;
  groqApiKeySet?: boolean;
  nvidiaApiKeySet?: boolean;
  selectedModel: string;
  selectedProvider: string;
  visionModel: string;
  visionProvider: string;
  wakeWord: string;
  voiceEnabled: boolean;
  selectedCharacterId: string;
  miniModeEnabled: boolean;
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  discordBotToken?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  notionApiKey?: string | null;
  /** @nullable */
  spotifyClientId?: string | null;
  /** @nullable */
  spotifyClientSecret?: string | null;
  /** @nullable */
  emailAddress?: string | null;
  /** @nullable */
  emailPassword?: string | null;
  /** @nullable */
  emailProvider?: string | null;
  googleClientIdSet?: boolean;
  googleClientSecretSet?: boolean;
  googleRefreshTokenSet?: boolean;
}

export interface SettingsInput {
  /** @nullable */
  groqApiKey?: string | null;
  /** @nullable */
  nvidiaApiKey?: string | null;
  /** @nullable */
  selectedModel?: string | null;
  /** @nullable */
  selectedProvider?: string | null;
  /** @nullable */
  visionModel?: string | null;
  /** @nullable */
  visionProvider?: string | null;
  /** @nullable */
  wakeWord?: string | null;
  /** @nullable */
  voiceEnabled?: boolean | null;
  /** @nullable */
  selectedCharacterId?: string | null;
  /** @nullable */
  miniModeEnabled?: boolean | null;
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  discordBotToken?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  notionApiKey?: string | null;
  /** @nullable */
  spotifyClientId?: string | null;
  /** @nullable */
  spotifyClientSecret?: string | null;
  /** @nullable */
  emailAddress?: string | null;
  /** @nullable */
  emailPassword?: string | null;
  /** @nullable */
  emailProvider?: string | null;
  /** @nullable */
  googleClientId?: string | null;
  /** @nullable */
  googleClientSecret?: string | null;
  /** @nullable */
  googleRefreshToken?: string | null;
}

export interface Message {
  id: number;
  role: string;
  content: string;
  createdAt: string;
  conversationId: number;
  /** @nullable */
  model?: string | null;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface ConversationInput {
  title: string;
}

export interface ConversationDetail {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export type CommandInputPayload = { [key: string]: unknown };

export interface CommandInput {
  /** open_app | search_web | set_reminder | get_time | get_weather | clipboard | screenshot | custom */
  type: string;
  payload: CommandInputPayload;
}

/**
 * @nullable
 */
export type CommandResultData = { [key: string]: unknown } | null;

export interface CommandResult {
  success: boolean;
  message: string;
  /** @nullable */
  data?: CommandResultData;
}

export interface CommandCategory {
  name: string;
  description: string;
  examples: string[];
}

export type StatsTopCommandsItem = {
  command: string;
  count: number;
};

export interface Stats {
  totalConversations: number;
  totalMessages: number;
  todayMessages: number;
  topCommands: StatsTopCommandsItem[];
}

export interface StopChatInput {
  /** The ID of the conversation to stop. */
  conversationId: number;
}