# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JARVIS AI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat:
    post:
      operationId: sendChat
      tags: [chat]
      summary: Send a message to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: AI service error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /settings:

    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current user settings (API keys masked)
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
  /chat/stop:
    post:
      operationId: stopChat
      tags: [chat]
      summary: Abort an ongoing chat execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopChatInput"
      responses:
        "200":
          description: Execution stopped successfully
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        conversationId:
          type: ["integer", "null"]
        model:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
          description: "groq | nvidia"
        imageBase64:
          type: ["string", "null"]
          description: "Optional base64 image data for multimodal vision models"
    ChatResponse:
      type: object
      required: [reply, conversationId, messageId]
      properties:
        reply:
          type: string
        conversationId:
          type: integer
        messageId:
          type: integer
        model:
          type: ["string", "null"]
        tokensUsed:
          type: ["integer", "null"]
        toolsUsed:
          type: ["array", "null"]
          items:
            type: string
...
    Settings:
      type: object
      required: [id, selectedModel, selectedProvider, visionModel, visionProvider, wakeWord, voiceEnabled, selectedCharacterId, miniModeEnabled]
      properties:
        id:
          type: integer
        groqApiKeySet:
          type: boolean
        nvidiaApiKeySet:
          type: boolean
        selectedModel:
          type: string
        selectedProvider:
          type: string
        visionModel:
          type: string
        visionProvider:
          type: string
        wakeWord:
          type: string
        voiceEnabled:
          type: boolean
        selectedCharacterId:
          type: string
        miniModeEnabled:
          type: boolean
        telegramBotToken:
          type: ["string", "null"]
        discordBotToken:
          type: ["string", "null"]
        systemPrompt:
          type: ["string", "null"]
        notionApiKey:
          type: ["string", "null"]
        spotifyClientId:
          type: ["string", "null"]
        spotifyClientSecret:
          type: ["string", "null"]
        emailAddress:
          type: ["string", "null"]
        emailPassword:
          type: ["string", "null"]
        emailProvider:
          type: ["string", "null"]
        googleClientIdSet:
          type: boolean
        googleClientSecretSet:
          type: boolean
        googleRefreshTokenSet:
          type: boolean
    SettingsInput:
      type: object
      properties:
        groqApiKey:
          type: ["string", "null"]
        nvidiaApiKey:
          type: ["string", "null"]
        selectedModel:
          type: ["string", "null"]
        selectedProvider:
          type: ["string", "null"]
        visionModel:
          type: ["string", "null"]
        visionProvider:
          type: ["string", "null"]
        wakeWord:
          type: ["string", "null"]
        voiceEnabled:
          type: ["boolean", "null"]
        selectedCharacterId:
          type: ["string", "null"]
        miniModeEnabled:
          type: ["boolean", "null"]
        telegramBotToken:
          type: ["string", "null"]
        discordBotToken:
          type: ["string", "null"]
        systemPrompt:
          type: ["string", "null"]
        notionApiKey:
          type: ["string", "null"]
        spotifyClientId:
          type: ["string", "null"]
        spotifyClientSecret:
          type: ["string", "null"]
        emailAddress:
          type: ["string", "null"]
        emailPassword:
          type: ["string", "null"]
        emailProvider:
          type: ["string", "null"]
        googleClientId:
          type: ["string", "null"]
        googleClientSecret:
          type: ["string", "null"]
        googleRefreshToken:
          type: ["string", "null"]
...
    StopChatInput:
      type: object
      required: [conversationId]
      properties:
        conversationId:
          type: integer
          description: The ID of the conversation to stop.