# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  groqApiKeySet?: boolean;
  nvidiaApiKeySet?: boolean;
  selectedModel: string;
  selectedProvider: string;
  visionModel: string;
  visionProvider: string;
  wakeWord: string;
  voiceEnabled: boolean;
  selectedCharacterId: string;
  miniModeEnabled: boolean;
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  discordBotToken?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  notionApiKey?: string | null;
  /** @nullable */
  spotifyClientId?: string | null;
  /** @nullable */
  spotifyClientSecret?: string | null;
  /** @nullable */
  emailAddress?: string | null;
  /** @nullable */
  emailPassword?: string | null;
  /** @nullable */
  emailProvider?: string | null;
  googleClientIdSet?: boolean;
  googleClientSecretSet?: boolean;
  googleRefreshTokenSet?: boolean;
}