# lib/api-zod/src/generated/types/settingsInput.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * JARVIS AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsInput {
  /** @nullable */
  groqApiKey?: string | null;
  /** @nullable */
  nvidiaApiKey?: string | null;
  /** @nullable */
  selectedModel?: string | null;
  /** @nullable */
  selectedProvider?: string | null;
  /** @nullable */
  visionModel?: string | null;
  /** @nullable */
  visionProvider?: string | null;
  /** @nullable */
  wakeWord?: string | null;
  /** @nullable */
  voiceEnabled?: boolean | null;
  /** @nullable */
  selectedCharacterId?: string | null;
  /** @nullable */
  miniModeEnabled?: boolean | null;
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  discordBotToken?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  notionApiKey?: string | null;
  /** @nullable */
  spotifyClientId?: string | null;
  /** @nullable */
  spotifyClientSecret?: string | null;
  /** @nullable */
  emailAddress?: string | null;
  /** @nullable */
  emailPassword?: string | null;
  /** @nullable */
  emailProvider?: string | null;
  /** @nullable */
  googleClientId?: string | null;
  /** @nullable */
  googleClientSecret?: string | null;
  /** @nullable */
  googleRefreshToken?: string | null;
}